Extracting Locabulary::Utility module
Given that I'm working at extracting Command objects, there is
duplication of logic that occurs throughout the code. I want to
extract that to a collaborating module that can be used elsewhere.
